add: docs

+++ include/keyboard_input.h
@@ -1,12 +1,26 @@
+/**
+ * @file keyboard_input.h
+ *
+ * @brief This file contains keyboard input functions.
+ *
+ * This file contains keyboard input functions.
+ */
+
 #ifndef KEYBOARD_INPUT_H
 #define KEYBOARD_INPUT_H
 
-#include "stdint.h"
 #include "stdio.h"
 #include "sys/ioctl.h"
 #include "termios.h"
 
-int16_t kbhit()
+/**
+ * @brief Check if a key is pressed.
+ *
+ * @brief It's an unblocking function that returns the number of bytes waiting in the input buffer.
+ *
+ * @return int The number of bytes waiting in the input buffer.
+ */
+int kbhit()
 {
     static const int STDIN = 0;
     static bool initialized = false;
